1.0.67 - Continue refactoring into finer-grained source modules. Fixed 'eslint' arguments to fully capture all files in 'src/'. Update Gallery to 0.0.85.
