+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "smartdown",
-  "version": "1.0.66",
+  "version": "1.0.67",
   "description": "A library for translating, rendering and interacting with Smartdown documents. Smartdown is an extension of Markdown that provides richer media support and reactive programming capability.",
   "main": "dist/lib/smartdown.js",
   "scripts": {
@@ -14,12 +14,12 @@
     "builddoc": "rm -rf distdoc/ distdocts/ && tsc --build doc/tsconfig_for_docs.json && BUILD=production node doc/build.js",
     "buildnoclean": "BUILD=production node_modules/.bin/webpack --progress",
     "build": "npm run clean && npm run builddoc && npm run buildnoclean",
-    "devbuild": "npm run build && (cd dist/ && ln -s . smartdown) && npm run serve",
+    "devbuild": "npm run buildnoclean && (mkdir devdist/ && mv dist devdist/smartdown/ && mv devdist dist) && npm run serve",
     "serve": "http-server --cors -p 4000 -c-1 -S -C ./ssl/cert.crt -K ./ssl/cert.key dist",
     "ghpublish": "./publish.sh",
     "watch": "BUILD=development node_modules/.bin/webpack --devtool=source-map --watch --progress",
     "dev": "(npm run serve &) && npm run watch",
-    "lint": "node_modules/.bin/eslint src/*.ts src/*.js src/extensions/*.js",
+    "lint": "node_modules/.bin/eslint 'src/**.js' 'src/**.ts'",
     "linttest": "node_modules/.bin/eslint test/*.js test/*/*.js",
     "test": "npm run test:unit && npm run test:e2e",
     "test:e2e": "node_modules/.bin/nightwatch",
@@ -47,11 +47,11 @@
   },
   "homepage": "https://github.com/smartdown/smartdown#readme",
   "devDependencies": {
-    "@babel/core": "^7.22.11",
+    "@babel/core": "^7.22.15",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-transform-modules-umd": "7.22.5",
-    "@babel/plugin-transform-runtime": "^7.22.10",
-    "@babel/preset-env": "^7.22.10",
+    "@babel/plugin-transform-runtime": "^7.22.15",
+    "@babel/preset-env": "^7.22.15",
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
     "autoprefixer": "^10.4.15",
@@ -85,13 +85,13 @@
     "ts-loader": "^9.4.4",
     "typescript": "4.9.5",
     "webpack": "^5.88.2",
-    "webpack-bundle-analyzer": "^4.9.0",
+    "webpack-bundle-analyzer": "^4.9.1",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1",
     "webpack-node-externals": "^3.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.22.11",
+    "@babel/runtime": "^7.22.15",
     "@stdlib/dist-datasets-sotu": "0.0.96",
     "@stdlib/dist-datasets-tree-exclude": "0.0.96",
     "@stdlib/dist-tree": "0.0.96",
@@ -110,7 +110,7 @@
     "lodash": "^4.17.21",
     "marked": "^3.0.4",
     "p5": "^1.7.0",
-    "smartdown-gallery": "0.0.84",
+    "smartdown-gallery": "0.0.85",
     "smoothscroll-polyfill": "^0.4.4",
     "stacktrace-js": "^2.0.2",
     "vdom-to-html": "^2.3.1",
