+++ src/parse/decodeInlineScript.js
