+++ src/render/entityEscape.js
