+++ src/util/areValuesSameEnough.js
