+++ test/mocha/b.configure.spec.js
@@ -17,8 +17,8 @@ describe('b.configure', function() {
   this.timeout(5000); // Default is 2000ms. Must be larger than postConfigureTimeout
 
   describe('smartdown.version()', function() {
-    it('should return "1.0.66"', function(done) {
-      assert.equal(smartdown.version, '1.0.66');
+    it('should return "1.0.67"', function(done) {
+      assert.equal(smartdown.version, '1.0.67');
       done();
     });
   });
